LayoutTests/fast/events/window-onerror17-expected.txt

Test that window.onerror is called on window object for events created using new ErrorEvent('error', {}).

Main frame window.onerror:
msg: Some error
url: Some file
line: 42
column: 43
error: Hello there
DONE.
